# chore(processors.unpivot): Cleanup code and improve performance