## Summary

This PR cleans up the plugin code and fixes naming of unit tests. As a side effect the PR improves performance by about 10%

```
goos: linux
goarch: amd64
pkg: github.com/influxdata/telegraf/plugins/processors/unpivot
cpu: AMD Ryzen 7 3700X 8-Core Processor   
```
| benchmark | before | after | improvement |
|--- | --- | --- | --- |
| BenchmarkAsTag-16 | 4561 ns/op |  4011 ns/op | **1.137x** |
| BenchmarkAsMetric-16 | 3598 ns/op |  3001 ns/op | **1.199x** |

## Checklist

- [x] No AI generated code was used in this PR

## Related issues